pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^4.0"
python = "^3.8"
django-model-utils = "^4.2.0"
black = "^22.10.0"
django-extensions = "^3.2.1"
tzdata = "^2022.5"
rich = "^12.6.0"
polyfile = "^0.4.2"
loguru = "^0.6.0"
[tool.poetry.dev-dependencies]
